+++ Shared/Foundation/Time/Time.h
@@ -2,7 +2,7 @@
 
 double GetTimeSeconds();
 double GetRawTimeSeconds();
-long GetRawTimeMilliseconds();
+long long GetRawTimeMilliseconds();
 
 void SetTimeScale(double scale);
 double GetTimeScale();
